Renaming graphic promo to graphic promotion everywhere
