+++ docs/src/constants/routes.js
@@ -97,7 +97,7 @@ export const INFINITE_SCROLL = '/components/infinite-scroll';
 export const BREADCRUMB = '/components/breadcrumb';
 export const FLARE = '/components/flare';
 export const BOTTOM_NAV = '/components/bottom-nav';
-export const GRAPHIC_PROMO = '/components/graphic-promo';
+export const GRAPHIC_PROMOTION = '/components/graphic-promotion';
 
 export const ALIGNMENT = '/components/alignment';
 export const THEMING = '/components/theming';
