--- docs/src/pages/components/GraphicPromo/WebGraphicPromo.mdx
@@ -1,22 +1,0 @@
----
-title: Graphic promo
-platform: web
-githubPath: bpk-component-graphic-promo
----
-
-
-import PresentationBlock from 'components/PresentationBlock';
-import Readme from 'bpk-component-graphic-promo/README.md';
-import { DefaultExample } from 'bpk-component-graphic-promo/examples';
-
-## Table of contents
-
-## Default
-
-<PresentationBlock>
-  <DefaultExample />
-</PresentationBlock>
-
-## Implementation
-
-<Readme />
