+++ docs/src/pages/components/GraphicPromotion/WebGraphicPromotion.mdx
@@ -1,0 +1,52 @@
+---
+title: Graphic promotion
+platform: web
+githubPath: bpk-component-graphic-promotion
+---
+
+
+import PresentationBlock from 'components/PresentationBlock';
+import Readme from 'bpk-component-graphic-promotion/README.md';
+import { DefaultExample } from 'bpk-component-graphic-promotion/examples';
+
+## Table of contents
+
+## Default
+
+<PresentationBlock>
+  <DefaultExample />
+</PresentationBlock>
+
+## Center Aligned Example
+
+<PresentationBlock>
+  <CenterAlignedExample />
+</PresentationBlock>
+
+## Right Aligned Example
+
+<PresentationBlock>
+  <RightAlignedExample />
+</PresentationBlock>
+
+## Inverted Portrait Example
+
+<PresentationBlock>
+  <InvertedPortraitExample />
+</PresentationBlock>
+
+## MinimalisticExample
+
+<PresentationBlock>
+  <MinimalisticExample />
+</PresentationBlock>
+
+## Non-Sponsored Example
+
+<PresentationBlock>
+  <NonSponsoredExample />
+</PresentationBlock>
+
+## Implementation
+
+<Readme />
