+++ package.json
@@ -75,8 +75,8 @@
     "@skyscanner/bpk-foundations-android": "^2.0.0",
     "@skyscanner/bpk-foundations-ios": "^1.1.0",
     "@skyscanner/bpk-foundations-react-native": "^1.1.0",
-    "@skyscanner/bpk-svgs": "^14.0.16",
     "@skyscanner/bpk-foundations-web": "^4.3.2",
+    "@skyscanner/bpk-svgs": "^14.0.16",
     "autoprefixer": "^9.8.6",
     "babel-jest": "^26.5.2",
     "babel-loader": "^8.0.6",
@@ -107,7 +107,7 @@
     "bpk-component-fieldset": "4.1.0",
     "bpk-component-flare": "2.1.0",
     "bpk-component-form-validation": "4.2.0",
-    "bpk-component-graphic-promo": "1.0.0",
+    "bpk-component-graphic-promotion": "^1.0.0",
     "bpk-component-grid": "3.1.0",
     "bpk-component-grid-toggle": "3.1.0",
     "bpk-component-horizontal-nav": "5.1.0",
